requirements.txt
beautifulsoup4==4.13.4
pandas==2.3.1
dbt-snowflake==1.10.0
sqlalchemy==2.0.43
python-dotenv>=0.9.9,<1.0.0
snowflake-sqlalchemy==1.7.6
google-generativeai==0.8.5
json5==0.12.1
google-genai==1.30.0
scikit-learn==1.7.1
